requirements.txt
web3
# rusty-rlp
uniswap-python

